implemented wishlist

+++ Controllers/WishlistController.cs
@@ -1,4 +1,6 @@
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using peakmotion.Models;
 using peakmotion.Repositories;
 using peakmotion.ViewModels;
 
@@ -7,34 +9,67 @@ namespace peakmotion.Controllers
     public class WishlistController : Controller
     {
         private readonly WishlistRepo _wishlistRepo;
+        private readonly PmuserRepo _pmuserRepo;
+        private readonly UserManager<IdentityUser> _userManager;
 
-        public WishlistController(WishlistRepo wishlistRepo)
+        public WishlistController(WishlistRepo wishlistRepo, PmuserRepo pmuserRepo, UserManager<IdentityUser> userManager)
         {
             _wishlistRepo = wishlistRepo;
+            _pmuserRepo = pmuserRepo;
+            _userManager = userManager;
         }
 
-        public IActionResult Index()
+        public async Task<IActionResult> Index()
         {
-            int userId = 1;
+            var identityUser = await _userManager.GetUserAsync(User);
+            var userId = _pmuserRepo.GetUserIdByUserEmail(identityUser.Email);
 
-            IEnumerable<WishlistVM> wishlistItems = _wishlistRepo.GetWishlistByUserId(userId);
-            return View(wishlistItems);
+            if (userId == null)
+            {
+                return RedirectToAction("Login", "Account");
+            }
+            else{
+                IEnumerable<WishlistVM> wishlistItems = _wishlistRepo.GetWishlistByUserId(userId);
+                return View(wishlistItems);
+            }
+
+
         }
 
         [HttpPost]
-        public IActionResult Add(int productId)
+        public async Task<IActionResult> Add(int productId)
         {
-            int userId = 1;
-            _wishlistRepo.AddToWishlist(userId, productId);
-            return RedirectToAction(nameof(Index));
+            var identityUser = await _userManager.GetUserAsync(User);
+            var userId = _pmuserRepo.GetUserIdByUserEmail(identityUser.Email);
+
+            if (userId == null)
+            {
+                return RedirectToAction("Login", "Account");
+            }
+            else
+            {
+                _wishlistRepo.AddToWishlist(userId, productId);
+                return RedirectToAction(nameof(Index));
+            }
+
         }
 
         [HttpPost]
-        public IActionResult Remove(int productId)
+        public async Task<IActionResult> Remove(int productId)
         {
-            int userId = 1;
-            _wishlistRepo.RemoveFromWishlist(userId, productId);
-            return RedirectToAction(nameof(Index));
+            var identityUser = await _userManager.GetUserAsync(User);
+            var userId = _pmuserRepo.GetUserIdByUserEmail(identityUser.Email);
+
+            if (userId == null)
+            {
+                return RedirectToAction("Login", "Account");
+            }
+            else
+            {
+                _wishlistRepo.RemoveFromWishlist(userId, productId);
+                return RedirectToAction(nameof(Index));
+            }
+
         }
     }
 }
